+++ georgePompidou/promo.php
@@ -4,6 +4,6 @@ namespace georgePompidou;
 
 interface promo
 {
-    function __construct(int $nombrePizza, int $gourmet = 0, int $prix = null, string $name, ?diametre $diametre = null);
+    function __construct(int $nombrePizza, int $gourmet = 0, float $prix = null, string $name, ?diametre $diametre = null);
     function checkValider(array $individuelles);
 }
